Made large builtins in the former std_pred.i into inlined subroutines. Added atom_codes, number_codes and made atom_chars, number_chars into synonyms to the *_codes builtins.

+++ emu/xsberror.c
@@ -68,7 +68,7 @@ void xsb_abort(char *description, ...)
     pcreg = exception_handler(message);
 
     /* this allows xsb_abort to jump out even from nested loops */
-    longjmp(xsb_abort_fallback_environment, pcreg);
+    longjmp(xsb_abort_fallback_environment, (int) pcreg);
 }
